+++ setup_app.py
@@ -4,7 +4,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="music-nas",
-    version="0.15.1-rc2",
+    version="0.15.1-rc5",
     packages=find_packages(),
     python_requires=">=3.10",
 )